# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexTripAI Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream Chat",
        "operationId": "stream_chat_api_chat_stream_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/preferences": {
      "get": {
        "tags": [
          "personalization"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_me_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizationProfile"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "personalization"
        ],
        "summary": "Reset Preferences",
        "operationId": "reset_preferences_api_me_preferences_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizationProfile"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "personalization"
        ],
        "summary": "Update Preferences",
        "operationId": "update_preferences_api_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizationProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/events": {
      "post": {
        "tags": [
          "personalization"
        ],
        "summary": "Record Preference Event",
        "operationId": "record_preference_event_api_me_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Record Preference Event Api Me Events Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "display_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Message"
          },
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "entity_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "kb_version": {
            "type": "string",
            "pattern": "^v[1-9][0-9]*$",
            "title": "Kb Version",
            "default": "v8"
          },
          "travel_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Date"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "include_weather": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Weather"
          }
        },
        "type": "object",
        "required": [
          "message",
          "session_id"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": "kb_retrieval"
          },
          "orchestration_mode": {
            "type": "string",
            "title": "Orchestration Mode",
            "default": "graph_only"
          },
          "resolved_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Resolved Context"
          },
          "kb_version": {
            "type": "string",
            "pattern": "^v[1-9][0-9]*$",
            "title": "Kb Version",
            "default": "v8"
          },
          "facts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Facts"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "itinerary": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Itinerary"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Trace"
          },
          "query_plan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query Plan"
          },
          "matched_paths": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Matched Paths"
          },
          "constraint_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Constraint Results"
          },
          "required_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Tools"
          },
          "clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Clarification"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherAssessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather_forecast": {
            "items": {
              "$ref": "#/components/schemas/WeatherAssessment"
            },
            "type": "array",
            "title": "Weather Forecast"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message_id",
          "answer"
        ],
        "title": "ChatResponse"
      },
      "Clarification": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "field": {
            "type": "string",
            "title": "Field"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/ClarificationOption"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "field"
        ],
        "title": "Clarification"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PersonalizationProfile": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          },
          "personalization_enabled": {
            "type": "boolean",
            "title": "Personalization Enabled",
            "default": true
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed",
            "default": false
          },
          "onboarding_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Onboarding Version",
            "default": 0
          },
          "profile_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Profile Revision",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "deletion_pending"
            ],
            "title": "Status",
            "default": "active"
          },
          "budget_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "budget",
                  "moderate",
                  "premium"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Level"
          },
          "travel_pace": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "relaxed",
                  "balanced",
                  "packed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Pace"
          },
          "party_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "solo",
                  "couple",
                  "family",
                  "friends",
                  "business"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Type"
          },
          "has_children": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Children"
          },
          "preferred_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Preferred Concepts"
          },
          "excluded_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Excluded Concepts"
          },
          "preferred_cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Preferred Cities"
          },
          "dietary_requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Dietary Requirements"
          },
          "accessibility_requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Accessibility Requirements"
          },
          "transport_preferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Transport Preferences"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "PersonalizationProfile"
      },
      "PersonalizationUpdate": {
        "properties": {
          "personalization_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personalization Enabled"
          },
          "onboarding_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed"
          },
          "onboarding_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Version"
          },
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "budget_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "budget",
                  "moderate",
                  "premium"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Level"
          },
          "travel_pace": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "relaxed",
                  "balanced",
                  "packed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Pace"
          },
          "party_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "solo",
                  "couple",
                  "family",
                  "friends",
                  "business"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Type"
          },
          "has_children": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Children"
          },
          "preferred_concepts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Concepts"
          },
          "excluded_concepts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Concepts"
          },
          "preferred_cities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Cities"
          },
          "dietary_requirements": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Dietary Requirements"
          },
          "accessibility_requirements": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessibility Requirements"
          },
          "transport_preferences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Transport Preferences"
          }
        },
        "type": "object",
        "title": "PersonalizationUpdate"
      },
      "PreferenceEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "like",
              "dislike",
              "save",
              "unsave",
              "add_to_itinerary",
              "remove_from_itinerary",
              "open_source"
            ],
            "title": "Event Type"
          },
          "place_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Place Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "place_id"
        ],
        "title": "PreferenceEvent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}